x64で型の警告をけす

+++ src/Caption2Ass_PCR_pf/Caption2Ass_PCR.cpp
@@ -1245,7 +1245,7 @@ static int main_loop(CAppHandler& app, CCaptionDllUtil& capUtil, CAPTION_LIST& c
   while (true){
 
     //ÉfÅ[É^ì«çû
-    int readNum;
+    size_t readNum;
     if (!afterResync)
     {
       //í èÌ
